+++ rustyhack_lib/src/message_handler/messages.rs
@@ -16,7 +16,7 @@ pub enum PlayerRequest {
     PlayerLogout(ClientDetails),
     UpdateVelocity(PositionMessage),
     PickupItem(PositionMessage),
-    DropItem(PositionMessage),
+    DropItem((u16, PositionMessage)),
     GetChunkedAllMaps,
     Timeout(String),
     Undefined,